@@ -332,7 +332,7 @@ target_sources(momo
 
 target_include_directories(momo PRIVATE src)
 
-set_target_properties(momo PROPERTIES CXX_STANDARD 17 C_STANDARD 99)
+set_target_properties(momo PROPERTIES CXX_STANDARD 20 C_STANDARD 99)
 
 target_link_libraries(momo
   PRIVATE
@@ -648,7 +648,7 @@ elseif (TARGET_OS STREQUAL "linux")
         NvCodec/NvCodec/NvDecoder/NvDecoder.cpp
         NvCodec/NvCodec/NvEncoder/NvEncoderCuda.cpp
       PROPERTIES
-        COMPILE_OPTIONS "-xcuda;--cuda-gpu-arch=sm_35;-std=gnu++14"
+        COMPILE_OPTIONS "-xcuda;--cuda-gpu-arch=sm_35;-std=gnu++20"
     )
 
     target_link_directories(momo PRIVATE /usr/local/cuda/lib64)
